+++ lending/src/sdk/loanwriteback.ts
@@ -7,6 +7,7 @@ import { LoanWritebackBankAccounts } from "./loanwritebackbankaccounts";
 import { LoanWritebackBankTransactions } from "./loanwritebackbanktransactions";
 import { LoanWritebackCreateOperations } from "./loanwritebackcreateoperations";
 import { LoanWritebackDirectCosts } from "./loanwritebackdirectcosts";
+import { LoanWritebackPayments } from "./loanwritebackpayments";
 import { LoanWritebackSuppliers } from "./loanwritebacksuppliers";
 import { LoanWritebackTransfers } from "./loanwritebacktransfers";
 import { SDKConfiguration } from "./sdk";
@@ -17,6 +18,7 @@ export class LoanWriteback {
     public bankTransactions: LoanWritebackBankTransactions;
     public createOperations: LoanWritebackCreateOperations;
     public directCosts: LoanWritebackDirectCosts;
+    public payments: LoanWritebackPayments;
     public suppliers: LoanWritebackSuppliers;
     public transfers: LoanWritebackTransfers;
     private sdkConfiguration: SDKConfiguration;
@@ -28,6 +30,7 @@ export class LoanWriteback {
         this.bankTransactions = new LoanWritebackBankTransactions(this.sdkConfiguration);
         this.createOperations = new LoanWritebackCreateOperations(this.sdkConfiguration);
         this.directCosts = new LoanWritebackDirectCosts(this.sdkConfiguration);
+        this.payments = new LoanWritebackPayments(this.sdkConfiguration);
         this.suppliers = new LoanWritebackSuppliers(this.sdkConfiguration);
         this.transfers = new LoanWritebackTransfers(this.sdkConfiguration);
     }