+++ lending/RELEASES.md
@@ -69,3 +69,13 @@ Based on:
 - [typescript v3.3.0] lending
 ### Releases
 - [NPM v3.3.0] https://www.npmjs.com/package/@codat/lending/v/3.3.0 - lending
+
+## 2023-10-16 01:28:13
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.99.2 (2.155.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.3.1] lending
+### Releases
+- [NPM v3.3.1] https://www.npmjs.com/package/@codat/lending/v/3.3.1 - lending